+++ taiga/projects/references/serializers.py
@@ -23,4 +23,16 @@ class ResolverSerializer(serializers.Serializer):
     us = serializers.IntegerField(required=False)
     task = serializers.IntegerField(required=False)
     issue = serializers.IntegerField(required=False)
+    ref = serializers.IntegerField(required=False)
     wikipage = serializers.CharField(max_length=512, required=False)
+
+    def validate(self, attrs):
+        if "ref" in attrs:
+            if "us" in attrs:
+                raise serializers.ValidationError("'us' param is incompatible with 'ref' in the same request")
+            if "task" in attrs:
+                raise serializers.ValidationError("'task' param is incompatible with 'ref' in the same request")
+            if "issue" in attrs:
+                raise serializers.ValidationError("'issue' param is incompatible with 'ref' in the same request")
+
+        return attrs